Moved `convertSVGsToImages()` to svgtoimageconverter.ts
